Boolean conversions added (only in if statements), along with returning values.
Returning values might have a type mismatch error. will be fixed in next commit. Check if.rsc in examples folder for new additions. If statements are fully fixed now, however do not support AND / OR operations. nested ifs work fine, but still no elif! working on it. __inbuilt__ renamed to extern, and a .md file talking about optimisations of redscript.